refactor(swap): fold Escape handling into CourseDropdownTrigger
The trigger already knows whether the dropdown is open, so it can own the
close affordances outright: clicking toggles, and Escape closes. Drops the
separate useEscapeToClose hook and the toggle logic each dropdown was
duplicating.
